pyproject.toml
[project]
name = "quant-pipeline"
version = "0.1.0"
description = "US equities data pipeline (Alpaca -> Parquet/DuckDB)"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.30.0",
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pytz>=2024.1",
]

[project.scripts]
qpipe = "quant_pipeline.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quant_pipeline"]
